requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2==2.9.10
python-dotenv==1.0.0
Werkzeug==2.3.7
